# web/src/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "PatchOperation": {
        "description": "One RFC 6902 operation. `move` and `copy` are rejected by the literal (400).",
        "properties": {
          "op": {
            "enum": [
              "add",
              "replace",
              "remove",
              "test"
            ],
            "title": "Op",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "op",
          "path"
        ],
        "title": "PatchOperation",
        "type": "object"
      },
      "ProblemDetails": {
        "description": "The one error body (RFC 7807 style).",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          },
          "errors": {
            "anyOf": [
              {
                "additionalProperties": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Errors"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "traceId": {
            "title": "Traceid",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "traceId"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
      "QueryField": {
        "description": "Input field definition. `type` is a free string; `multiple` and `match` have extra rules.",
        "properties": {
          "from": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From"
          },
          "regex": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regex"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "type"
        ],
        "title": "QueryField",
        "type": "object"
      },
      "QueryParam": {
        "description": "Template parameter. `type` is a free string; `values` is required when `type == \"array\"`.",
        "properties": {
          "default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "hint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hint"
          },
          "multiple": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Multiple"
          },
          "optional": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Optional"
          },
          "type": {
            "title": "Type",
            "type": "string"
          },
          "values": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          }
        },
        "required": [
          "type"
        ],
        "title": "QueryParam",
        "type": "object"
      },
...
      "QueryTemplate": {
        "description": "Stored/returned template. Also the target of PATCH re-validation.",
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "columnId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columnid"
          },
          "columns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "createdBy": {
            "title": "Createdby",
            "type": "string"
          },
          "database": {
            "title": "Database",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryField"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "isDeleted": {
            "default": false,
            "title": "Isdeleted",
            "type": "boolean"
          },
          "isManaged": {
            "default": false,
            "title": "Ismanaged",
            "type": "boolean"
          },
          "menu": {
            "title": "Menu",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryParam"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "queryType": {
            "$ref": "#/components/schemas/QueryType"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "updated": {
            "title": "Updated",
            "type": "string"
          },
          "updatedBy": {
            "title": "Updatedby",
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "queryType",
          "menu",
          "summary",
          "path",
          "cluster",
          "database",
          "query",
          "uuid",
          "updated",
          "createdBy",
          "updatedBy"
        ],
        "title": "QueryTemplate",
        "type": "object"
      },
      "QueryTemplateCreate": {
        "description": "POST body: audit fields (`createdBy`, `updatedBy`, `updated`) are ignored if sent.",
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "columnId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columnid"
          },
          "columns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "database": {
            "title": "Database",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryField"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "isDeleted": {
            "default": false,
            "title": "Isdeleted",
            "type": "boolean"
          },
          "isManaged": {
            "default": false,
            "title": "Ismanaged",
            "type": "boolean"
          },
          "menu": {
            "title": "Menu",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryParam"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "queryType": {
            "$ref": "#/components/schemas/QueryType"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "queryType",
          "menu",
          "summary",
          "path",
          "cluster",
          "database",
          "query",
          "uuid"
        ],
        "title": "QueryTemplateCreate",
        "type": "object"
      },
      "QueryTemplateReplace": {
        "description": "PUT body: `uuid` must match the path; `isDeleted` and audit fields are not accepted.",
        "properties": {
          "cluster": {
            "title": "Cluster",
            "type": "string"
          },
          "columnId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columnid"
          },
          "columns": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Columns"
          },
          "database": {
            "title": "Database",
            "type": "string"
          },
          "fields": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryField"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fields"
          },
          "isManaged": {
            "default": false,
            "title": "Ismanaged",
            "type": "boolean"
          },
          "menu": {
            "title": "Menu",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/QueryParam"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Params"
          },
          "path": {
            "items": {
              "type": "string"
            },
            "title": "Path",
            "type": "array"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "queryType": {
            "$ref": "#/components/schemas/QueryType"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "uuid": {
            "format": "uuid",
            "title": "Uuid",
            "type": "string"
          }
        },
        "required": [
          "name",
          "queryType",
          "menu",
          "summary",
          "path",
          "cluster",
          "database",
          "query",
          "uuid"
        ],
        "title": "QueryTemplateReplace",
        "type": "object"
      },
      "QueryType": {
        "enum": [
          "view",
          "query"
        ],
        "title": "QueryType",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "TIM API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/templates/queries": {
      "get": {
        "operationId": "listTemplates",
        "parameters": [
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "includeDeleted",
            "required": false,
            "schema": {
              "default": false,
              "title": "Includedeleted",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QueryTemplate"
                  },
                  "title": "Response List Templates Api Templates Queries Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "List Templates",
        "tags": [
          "templates"
        ]
      },
      "post": {
        "operationId": "createTemplate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTemplateCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTemplate"
                }
              }
            },
            "description": "Successful Response",
            "headers": {
              "Location": {
                "description": "`/api/templates/queries/{uuid}`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Already exists (`conflict`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create Template",
        "tags": [
          "templates"
        ]
      }
    },
    "/api/templates/queries/{uuid}": {
      "delete": {
        "operationId": "deleteTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found (`not-found`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Delete Template",
        "tags": [
          "templates"
        ]
      },
      "get": {
        "operationId": "getTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found (`not-found`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Get Template",
        "tags": [
          "templates"
        ]
      },
      "patch": {
        "operationId": "patchTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "items": {
                  "$ref": "#/components/schemas/PatchOperation"
                },
                "minItems": 1,
                "title": "Operations",
                "type": "array"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found (`not-found`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Patch Template",
        "tags": [
          "templates"
        ]
      },
      "put": {
        "operationId": "replaceTemplate",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueryTemplateReplace"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryTemplate"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed, or the cluster is not allowed (`validation` / `cluster-not-allowed`)."
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Missing, invalid or expired token (`unauthorized`).",
            "headers": {
              "WWW-Authenticate": {
                "description": "`Bearer`",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found (`not-found`)."
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unhandled error (`internal`)."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Replace Template",
        "tags": [
          "templates"
        ]
      }
    }
  }
}